fix(gen): Fix jsformat error and add back .gitignore for generated apps
Be careful to remove yuour local sublime text jsFormat settings as we use .jsbeautifyrc files now. This commit fixes some formatting that screwed up because I didn't remove the local settings. This was in the Formatter and Groupt utilities used by the Fiori apps.
Fixed silly bug introduced when refactoring the generator. .gitignore is now scaffolded for all genrated apps.