fix(tiles app): Fix the Tiles app to use the correct port if the user doesn't use the mock northwind service.
When a Tiles app was genreated the url for using the real Northwind service still had 8080 hard-coded as the port - fixed to be the localServerPort variable.
Improved the comments in the Tile app index.html regarding the mock OData service. Explaining that the two scripts enabling the mock service can be deleted to use the real Northwind service and that the proxy must then be configured in Gruntfile.js to get around CORS issue.