refactor(fiori apps): Change to EventBus for navigation
Refactor the fiori-style apps (Splitter, Tiles) to use the EventBus for navigation instead of coupling all the page view controllers to the root controller.
Whether this really simplifies your app or not is a matter of personal preference, but it is the best-practice according to the current UI5 SDK: https://openui5.hana.ondemand.com/#docs/guide/BestPractice.html.

Also includes minor adjustments to standardise the use of whitespace.

Fixes #50